Remove title-annotation endpoint and move it into a function
